+++ packages/falcon-shop-data/src/Customer/ChangePasswordMutation.tsx
@@ -1,8 +1,9 @@
 import gql from 'graphql-tag';
+import { useMutation, MutationHookOptions } from '@apollo/react-hooks';
 import { Mutation, OperationInput } from '@deity/falcon-data';
 import { ChangePasswordInput } from '@deity/falcon-shop-extension';
 
-export const CHANGE_PASSWORD = gql`
+export const CHANGE_PASSWORD_MUTATION = gql`
   mutation changePassword($input: ChangePasswordInput!) {
     changePassword(input: $input)
   }
@@ -14,6 +15,10 @@ export type ChangePasswordResponse = {
 
 export class ChangePasswordMutation extends Mutation<ChangePasswordResponse, OperationInput<ChangePasswordInput>> {
   static defaultProps = {
-    mutation: CHANGE_PASSWORD
+    mutation: CHANGE_PASSWORD_MUTATION
   };
 }
+
+export const useChangePasswordMutation = (
+  options: MutationHookOptions<ChangePasswordResponse, OperationInput<ChangePasswordInput>> = {}
+) => useMutation(CHANGE_PASSWORD_MUTATION, options);